// src/app/models/funcionario/funcionario-response.ts
export interface FuncionarioResponse{
    id: number;
    nome: string;
    documento: string;
    email: string;
    cargo: string;
    criadoEm: Date;
    alteradoEm: Date;
    status: string;
    username: string;
    tipoAutorizacao: string;
    statusFuncionario: string;
    areaAtuacao: string;
    salario: number;
    dataAdmissao: Date; 
}